feat(237-02): extract search facets and collections
- move facet count CTE/query logic into service_facets.py

- move collection search and visible member counts into service_collections.py

- keep get_facet_counts and search_collections re-exported from the public service facade
