+++ backend/app/modules/catalog/search/service.py
@@ -10,17 +10,7 @@ if TYPE_CHECKING:
     from app.platform.storage.provider import StorageProvider
 
 import structlog
-from sqlalchemy import (
-    String as SAString,
-    Select,
-    case,
-    collate,
-    func,
-    literal,
-    or_,
-    select,
-    text,
-)
+from sqlalchemy import Select, case, collate, func, literal, select, text
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
 from sqlalchemy.sql.elements import ColumnElement, Label
@@ -29,18 +19,18 @@ from app.core.identity import Identity
 from app.modules.auth.models import User
 from app.core.config import settings
 from app.modules.catalog.authorization import apply_visibility_filter
-from app.modules.catalog.collections.models import Collection, CollectionDataset
 from app.modules.catalog.datasets.domain.models import (
     Dataset,
     DatasetGrant,
     Record,
-    RecordKeyword,
 )
 from app.modules.catalog.datasets.domain.utils import extract_bbox
 from app.standards.ogc.utils import build_url
 from app.core.persistent_config import EMBEDDING_MODEL, SEMANTIC_SEARCH_ENABLED
 from app.modules.catalog.sources.provenance import derive_last_edited
 from app.platform.extensions import get_catalog_port
+from app.modules.catalog.search.service_collections import search_collections
+from app.modules.catalog.search.service_facets import get_facet_counts
 from app.modules.catalog.search.service_filters import (
     FacetCounts,
     SearchFilters,
@@ -197,195 +187,6 @@ def _compute_rrf_scores(
 
     # Sort by RRF score descending
     return sorted(scores.keys(), key=lambda rid: scores[rid], reverse=True)
-
-
-async def get_facet_counts(
-    session: AsyncSession,
-    user: Identity | None,
-    user_roles: set[str],
-    filters: SearchFilters,
-) -> FacetCounts:
-    """Return multi-group facet counts for datasets matching the given filters.
-
-    Returns dict with keys: record_type, keywords, source_organization, srid.
-    Does NOT filter by record_type itself (facets show counts for all types).
-    Separately counts matching collections from the collections table.
-    """
-    # Build a CTE that materializes the filtered (dataset_id, record_id) pairs
-    # once. Both the record_type counts and per-facet queries join against it
-    # instead of each independently re-evaluating the full filter stack.
-    filtered_base = (
-        select(Dataset.id.label("dataset_id"), Record.id.label("record_id"))
-        .select_from(Dataset)
-        .join(Record, Dataset.record_id == Record.id)
-    )
-    filtered_base = apply_visibility_filter(
-        filtered_base, user, user_roles, Record, DatasetGrant
-    )
-    filtered_base = _apply_common_filters(filtered_base, filters)
-    filtered_cte = filtered_base.cte("filtered_ids")
-
-    # Record-type counts from the CTE (replaces duplicate filter stack)
-    type_stmt = (
-        select(Record.record_type, func.count().label("count"))
-        .select_from(filtered_cte)
-        .join(Record, Record.id == filtered_cte.c.record_id)
-        .group_by(Record.record_type)
-    )
-    result = await session.execute(type_stmt)
-    counts = {row.record_type: row.count for row in result.all()}
-
-    # Separately count collections from the collections table
-    coll_stmt = select(func.count()).select_from(Collection)
-    if filters.q and filters.q.strip():
-        q_like = f"%{filters.q.strip().lower()}%"
-        coll_stmt = coll_stmt.where(
-            or_(
-                func.lower(Collection.name).like(q_like),
-                func.lower(func.coalesce(Collection.description, "")).like(q_like),
-            )
-        )
-    coll_count = (await session.execute(coll_stmt)).scalar_one()
-    if coll_count > 0:
-        counts["collection"] = coll_count
-
-    # Facet queries are intentionally sequential — SQLAlchemy AsyncSession
-    # is not safe for concurrent execute() on a shared connection.
-    # Each query joins the CTE instead of re-applying all filters.
-
-    # --- Keyword facets (top 20) ---
-    kw_stmt = (
-        select(RecordKeyword.keyword, func.count().label("count"))
-        .select_from(filtered_cte)
-        .join(RecordKeyword, RecordKeyword.record_id == filtered_cte.c.record_id)
-        .group_by(RecordKeyword.keyword)
-        .order_by(func.count().desc())
-        .limit(20)
-    )
-    kw_result = await session.execute(kw_stmt)
-    keyword_facets = [
-        {"value": row.keyword, "count": row.count} for row in kw_result.all()
-    ]
-
-    # --- Source organization facets ---
-    org_stmt = (
-        select(Record.source_organization, func.count().label("count"))
-        .select_from(filtered_cte)
-        .join(Record, Record.id == filtered_cte.c.record_id)
-        .where(Record.source_organization.isnot(None), Record.source_organization != "")
-        .group_by(Record.source_organization)
-        .order_by(func.count().desc())
-        .limit(50)
-    )
-    org_result = await session.execute(org_stmt)
-    org_facets = [
-        {"value": row.source_organization, "count": row.count}
-        for row in org_result.all()
-    ]
-
-    # --- SRID facets ---
-    srid_stmt = (
-        select(
-            func.cast(Dataset.srid, SAString).label("srid_str"),
-            func.count().label("count"),
-        )
-        .select_from(filtered_cte)
-        .join(Dataset, Dataset.id == filtered_cte.c.dataset_id)
-        .where(Dataset.srid.isnot(None))
-        .group_by(Dataset.srid)
-        .order_by(func.count().desc())
-        .limit(50)
-    )
-    srid_result = await session.execute(srid_stmt)
-    srid_facets = [
-        {"value": row.srid_str, "count": row.count} for row in srid_result.all()
-    ]
-
-    # --- Collections facet (lightweight: id, name, visible member count) ---
-    coll_facet_stmt = (
-        select(
-            Collection.id,
-            Collection.name,
-            func.count(CollectionDataset.dataset_id).label("dataset_count"),
-        )
-        .select_from(Collection)
-        .join(CollectionDataset, CollectionDataset.collection_id == Collection.id)
-        .join(filtered_cte, filtered_cte.c.dataset_id == CollectionDataset.dataset_id)
-        .group_by(Collection.id, Collection.name)
-        .order_by(func.count(CollectionDataset.dataset_id).desc())
-    )
-    coll_facet_result = await session.execute(coll_facet_stmt)
-    collections_facet = [
-        {"id": str(row.id), "name": row.name, "dataset_count": row.dataset_count}
-        for row in coll_facet_result.all()
-    ]
-
-    return {
-        "record_type": counts,
-        "keywords": keyword_facets,
-        "source_organization": org_facets,
-        "srid": srid_facets,
-        "collections": collections_facet,
-    }
-
-
-async def search_collections(
-    session: AsyncSession,
-    q: str,
-    user: Identity | None,
-    user_roles: set[str],
-    *,
-    limit: int = 10,
-) -> list[dict]:
-    """Search collections by text and return with visible member counts.
-
-    When q is empty, returns all collections (up to limit).
-    """
-    coll_stmt = select(Collection).limit(limit)
-
-    if q and q.strip():
-        q_like = f"%{q.strip().lower()}%"
-        coll_stmt = coll_stmt.where(
-            or_(
-                func.lower(Collection.name).like(q_like),
-                func.lower(func.coalesce(Collection.description, "")).like(q_like),
-            )
-        )
-    coll_result = await session.execute(coll_stmt)
-    collections = coll_result.scalars().all()
-
-    if not collections:
-        return []
-
-    # Get visible member counts in a single query
-    coll_ids = [c.id for c in collections]
-    member_stmt = (
-        select(
-            CollectionDataset.collection_id,
-            func.count().label("cnt"),
-        )
-        .select_from(CollectionDataset)
-        .join(Dataset, CollectionDataset.dataset_id == Dataset.id)
-        .join(Record, Dataset.record_id == Record.id)
-        .where(CollectionDataset.collection_id.in_(coll_ids))
-    )
-    member_stmt = apply_visibility_filter(
-        member_stmt, user, user_roles, Record, DatasetGrant
-    )
-    member_stmt = member_stmt.group_by(CollectionDataset.collection_id)
-    member_result = await session.execute(member_stmt)
-    count_map = {row.collection_id: row.cnt for row in member_result.all()}
-
-    return [
-        {
-            "id": str(c.id),
-            "name": c.name,
-            "description": c.description,
-            "dataset_count": count_map.get(c.id, 0),
-            "created_at": c.created_at.isoformat(),
-        }
-        for c in collections
-    ]
 
 
 def _build_fts_rank_col(
